// datasette_comments/frontend/src/page_data/ContentScriptPageData.types.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type ViewName = string;
export type Database = string | null;
export type Table = string | null;
export type ActorId = string;
export type Name = string | null;
export type ProfilePhotoUrl = string | null;
export type Username = string | null;
export type ReadonlyViewer = boolean;

export interface ContentScriptPageData {
  view_name: ViewName;
  database?: Database;
  table?: Table;
  author: Author;
  readonly_viewer: ReadonlyViewer;
  [k: string]: unknown;
}
export interface Author {
  actor_id: ActorId;
  name?: Name;
  profile_photo_url?: ProfilePhotoUrl;
  username?: Username;
  [k: string]: unknown;
}